refactor: user guards on a single file
